Chore: Refactor the interval state from the state sync
