+++ sqlmesh/core/state_sync/engine_adapter/__init__.py
@@ -1,0 +1,3 @@
+from sqlmesh.core.state_sync.engine_adapter.facade import EngineAdapterStateSync
+
+__all__ = ["EngineAdapterStateSync"]
